change policies folder struture
